policiesss

+++ app/Http/Controllers/Admin/UserController.php
@@ -33,11 +33,7 @@ class UserController extends Controller
     }
 
     public function index(Request $request) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
-
+        Gate::authorize("viewAny", User::class);
         $users = $this->filterUsers(User::query(), $request);
         return view("admin.users.index", [
             "users" => $users
@@ -47,19 +43,12 @@ class UserController extends Controller
     }
 
     public function create(Request $request) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
-
+        Gate::authorize("create", User::class);
         return view("admin.users.create");
     }
 
     public function store(Request $request) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
+        Gate::authorize("create", User::class);
         $request->validate([
             "name" => "required|max:255",
             "email" => "required|email|max:255|unique:users,email",
@@ -80,30 +69,21 @@ class UserController extends Controller
     }
 
     public function show(Request $request, User $user) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
+        Gate::authorize("view", $user);
         return view("admin.users.show", [
             "user" => $user
         ]);
     }
 
     public function edit(Request $request, User $user) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
+        Gate::authorize("update", $user);
         return view("admin.users.edit", [
             "user" => $user
         ]);
     }
 
     public function update(Request $request, User $user) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
+        Gate::authorize("update", $user);
         $data = $request->validate([
             "name" => "required|max:255",
             "email" => "required|email|max:255",
@@ -120,17 +100,14 @@ class UserController extends Controller
     }
 
     public function delete(Request $request, User $user) {
-        Gate::allowIf(
-            $request->user() &&
-            $request->user()->role === "admin"
-        );
-
+        Gate::authorize("delete", $user);
         return view("admin.users.delete", [
             "user" => $user
         ]);
     }
 
     public function destroy(User $user) {
+        Gate::authorize("delete", $user);
         $user->delete();
         return redirect()->route("admin.users.index")->with([
             "success" => "User deleted!"
@@ -138,6 +115,7 @@ class UserController extends Controller
     }
 
     public function trash(Request $request) {
+        Gate::authorize("viewAny", User::class);
         $users = $this->filterUsers(User::onlyTrashed(), $request);
         return view("admin.users.trash", [
             "users" => $users
@@ -147,6 +125,7 @@ class UserController extends Controller
     }
 
     public function restore(User $user) {
+        Gate::authorize("restore", $user);
         if (!$user->deleted_at)
             return back()->withErrors([
                 "user" => "Not found"
